requirements.txt
ultralytics
opencv-python
tensorflow
Pillow